requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==2.2.16
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==4.0.1
idna==3.3
importlib-metadata==1.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.3
Pillow==8.3.1
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0